% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCO2sim.within4.R
\name{VCO2_sim.within}
\alias{VCO2_sim.within}
\title{VCO2_sim.within Determines the Probability that Two VCO2 Measures are the Same Within Days or Within an Exercise Trial}
\usage{
VCO2_sim.within(a, b, plot = F)
}
\arguments{
\item{a}{The first VCO2 value obtained}

\item{b}{The second VCO2 value obtained}

\item{plot}{False returns the probability they are same distribution. True returns a plot of overlapping distributions.}
}
\description{
This function models the reliability of VCO2 data from the Parvomedics 2400 TrueOne system as a truncated
univariate normal distribution based on the work by Macfarlane and Wu 2013
Additional credit goes to Stackexchange user Wolfgang for his example code of the overlapping coefficient
}
\examples{
put something here
}
