% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VEsim3.R
\name{VE_sim}
\alias{VE_sim}
\title{VEsim Determines the Probability that Two VE Measures are the Same}
\usage{
VE_sim(a, b, system_a = "parvo_2400", system_b = "parvo_2400", plot = F)
}
\arguments{
\item{a}{The first VE value obtained}

\item{b}{The second VE value obtained}

\item{system_a}{The system used, in quotes, to obtain measurement 'a'.
Accepts "parvo_2400" and "douglas_bag". Defaults to parvo_2400.}

\item{system_b}{The system used, in quotes, to obtain measurement 'b'.
Accepts "parvo_2400" and "douglas_bag". Defaults to parvo_2400.}

\item{plot}{False returns the probability they are same distribution. True returns a plot of overlapping distributions.}
}
\description{
This function models the reliability of VE data from a particular system as univariate normal
based on the work by Crouter 2006
Additional credit goes to Stackexchange user Wolfgang for his example code of the overlapping coefficient
}
\examples{
put something here
}

