% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEsim.within4.R
\name{VE_sim.within}
\alias{VE_sim.within}
\title{VE_sim.within Determines the Probability that Two VE Measures are the Same Within Days or Within an Exercise Trial}
\usage{
VE_sim.within(a, b, plot = F)
}
\arguments{
\item{a}{The first VE value obtained}

\item{b}{The second VE value obtained}

\item{plot}{False returns the probability they are same distribution. True returns a plot of overlapping distributions.}
}
\description{
This function models the reliability of VE data from the Parvomedics 2400 TrueOne system as a truncated
univariate normal distribution based on the work by Macfarlane and Wu 2013
Additional credit goes to Stackexchange user Wolfgang for his example code of the overlapping coefficient
}
\examples{
put something here
}
