% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VO2sim4.R
\name{VO2_sim}
\alias{VO2_sim}
\title{VO2sim Determines the Probability that Two VO2 Measures are the Same Between Testing Days}
\usage{
VO2_sim(a, b, system_a = "parvo_2400", system_b = "parvo_2400", plot = F)
}
\arguments{
\item{a}{The first VO2 value obtained}

\item{b}{The second VO2 value obtained}

\item{system_a}{The system used, in quotes, to obtain measurement 'a'.
Accepts "parvo_2400" and "douglas_bag". Defaults to parvo_2400.}

\item{system_b}{The system used, in quotes, to obtain measurement 'b'.
Accepts "parvo_2400" and "douglas_bag". Defaults to parvo_2400.}

\item{plot}{False returns the probability they are same distribution. True returns a plot of overlapping distributions.}
}
\description{
This function models the reliability of VO2 data from a particular system as univariate normal
based on the work by Crouter 2006
Additional credit goes to Stackexchange user Wolfgang for his example code of the overlapping coefficient
}
\examples{
put something here
}
