% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VO2sim.within4.R
\name{VO2_sim.within}
\alias{VO2_sim.within}
\title{VO2_sim.within Determines the Probability that Two VO2 Measures are the Same Within Days or Within an Exercise Trial}
\usage{
VO2_sim.within(a, b, plot = F)
}
\arguments{
\item{a}{The first VO2 value obtained}

\item{b}{The second VO2 value obtained}

\item{plot}{False returns the probability they are same distribution. True returns a plot of overlapping distributions.}
}
\description{
This function models the reliability of VO2 data from the Parvomedics 2400 TrueOne system as a truncated
univariate normal distribution based on the work by Macfarlane and Wu 2013
Additional credit goes to Stackexchange user Wolfgang for his example code of the overlapping coefficient
}
\examples{
put something here
}
